Let EenyMeeny::Middleware clean up cookies from old (unconfigured) experiments.
Unless experiments are configured with an 'end_at' time then the experiment cookie lives for 1 month.
It is desireable to clean up "old experiement" cookies to avoid the HTTP COOKIE header being bloated with unused data.
This is especially true if many experiments are being run for shorter periods of time.